added custom exceptions
